✨ feat(register): Agregar combo en la vista de registro y fusionar manejadores
Se agregó un combo en la vista de registro (`Registrar.cshtml`) para mejorar la experiencia del usuario.

Se fusionaron los manejadores de usuarios eliminando `ManejadorUsuarios.cs` y consolidando la lógica en `ManejadorUsuario.cs`. Se realizaron ajustes en `UsuariosController.cs` y `Persona.cs` para reflejar estos cambios.

Otros cambios:
- Se actualizó la conexión en `Connection.cs`.
- Se añadieron archivos relacionados con `TipoDocumento` (`ManejadorTipoDocumento.cs` y `TipoDocumento.cs`).
