+++ BusinessLogic/ManejadorUsuario.cs
@@ -42,6 +42,8 @@ namespace BusinessLogic
             return conexion.EjecutarConsulta("login_usuario", parametros);
         }
 
+
+        /*Revisar por favor*/
         public List<Usuario> ObtenerUsuarios(int id)
         {
             List<Parametro> parametros = new List<Parametro>()
@@ -68,6 +70,34 @@ namespace BusinessLogic
                         Documento = Convert.ToInt32(row["DOCUMENTO_IDENTIDAD"].ToString()),
                         Telefono = row["TELEFONO"].ToString()
                     }
+                });
+            }
+            return usuarios;
+        }
+
+        public List<Usuario> ConsultarUsuarios()
+        {
+            DataTable datos = conexion.EjecutarConsulta("consultar_usuario");
+            List<Usuario> usuarios = new List<Usuario>();
+
+            foreach (DataRow row in datos.Rows)
+            {
+                usuarios.Add(new Usuario
+                {
+                    Id = Convert.ToInt32(row["PK_ID_USUARIO"]),
+                    Persona = new Persona
+                    {
+                        Nombre = row["NOMBRES"].ToString(),
+                        Apellido = row["APELLIDOS"].ToString(),
+                        Telefono = row["TELEFONO"].ToString(),
+                        Estado = Convert.ToInt32(row["ESTADO"]) == 1 ? "Activo" : "Inactivo"
+                    },
+                    Rol = new Rol
+                    {
+                        Nombre = row["NOMBRE_ROL"].ToString()
+                    },
+                    Correo = row["NOMBRE_USUARIO"].ToString(),
+                    CodigoReferido = 0,
                 });
             }
             return usuarios;
@@ -198,16 +228,35 @@ namespace BusinessLogic
         public bool CambiarContrasenia(int userId, string nuevaContrasenia)
         {
             List<Parametro> parametros = new List<Parametro>
-    {
-        new Parametro("p_id_usuario", userId),
-        new Parametro("p_nueva_contrasenia", Seguridad.Encriptar(nuevaContrasenia))
-    };
+            {
+                new Parametro("p_id_usuario", userId),
+                new Parametro("p_nueva_contrasenia", Seguridad.Encriptar(nuevaContrasenia))
+            };
 
             return conexion.EjecutarTransaccion("recuperar_contrasenia", parametros);
         }
 
 
 
+
+        public bool EditarEstadoUsuario(int idUsuario, int nuevoEstado)
+        {
+            List<Parametro> parametros = new List<Parametro>
+            {
+                new Parametro("p_id_usuario", idUsuario),
+                new Parametro("p_estado", nuevoEstado)
+            };
+
+            try
+            {
+                bool resultado = conexion.EjecutarTransaccion("editar_estado_usuario", parametros);
+                return resultado;
+            }
+            catch (Exception ex)
+            {
+                return false;
+            }
+        }
 
     }
 }
