+++ BusinessLogic/Models/TipoDocumento.cs
@@ -1,0 +1,14 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace BusinessLogic.Models
+{
+    public class TipoDocumento
+    {
+        public int Id { get; set; }
+        public string Nombre { get; set; }
+    }
+}
