+++ DataAccess/Connection.cs
@@ -10,7 +10,7 @@ namespace DataAccess
         public MySqlConnection connection;
         public bool Conectar()
         {
-            string cadenaConnection = "server=localhost; database=abastecete; user=root; password=smati0422; port=3306";
+            string cadenaConnection = "server=localhost; database=abastecete; user=root; password=0518; port=3306";
             connection = new MySqlConnection(cadenaConnection);
             try
             {
