+++ package.json
@@ -7,9 +7,8 @@
     "build:css": "tailwindcss build styles.css -o output.css"
   },
   "devDependencies": {
-    "tailwindcss": "^3.0.0",
-    "postcss": "^8.0.0",
-    "autoprefixer": "^10.0.0"
-  },
-  "dependencies": {}
+    "autoprefixer": "^10.4.20",
+    "postcss": "^8.5.3",
+    "tailwindcss": "^3.4.17"
+  }
 }
